Implement logging in WADL-tools.
XSL messages can now be directed to standard logs.  Test, use log4j.
A message should start with one of the items below:

[TRACE]
[DEBUG]
[INFO]
[WARNING]
[ERROR]

If it does then it will be sent the the appropriate log level.  If no
level is specified at the beginning of the line then WARNING is
assumed.